FIX: CL-2577 Field property "size" is only of type 'short' (i.e. max 32768) which is limiting e.g. for byte fields
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk-3-4@13146 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/metadata/DataRecordMetadata.java
@@ -1309,7 +1309,7 @@ public class DataRecordMetadata implements Serializable, Iterable<DataFieldMetad
 	public String getFieldNamesHeader(String[] excludedFieldNames, boolean quotedString, Character quoteChar) {
 		StringBuilder ret = new StringBuilder();
 		DataFieldMetadata dataFieldMetadata;
-		short fieldSize;
+		int fieldSize;
 		int headerSize;
 		char blank = ' ';
 		String label;
